GraalVM native image related fix to support governance-aggr and adapot

+++ aggregates/governance-aggr/src/main/java/com/bloxbean/cardano/yaci/store/governanceaggr/processor/DRepStatusProcessor.java
@@ -6,6 +6,7 @@ import com.bloxbean.cardano.yaci.store.events.RollbackEvent;
 import com.bloxbean.cardano.yaci.store.events.internal.PreCommitEvent;
 import com.bloxbean.cardano.yaci.store.governance.domain.DRepRegistration;
 import com.bloxbean.cardano.yaci.store.governance.domain.event.DRepRegistrationEvent;
+import com.bloxbean.cardano.yaci.store.governanceaggr.GovernanceAggrProperties;
 import com.bloxbean.cardano.yaci.store.governanceaggr.domain.DRep;
 import com.bloxbean.cardano.yaci.store.governanceaggr.storage.DRepStorage;
 import com.bloxbean.cardano.yaci.store.governanceaggr.storage.impl.model.DRepStatus;
@@ -25,18 +26,25 @@ import java.util.List;
 @Slf4j
 // todo: 'inactive', 'active again' case
 public class DRepStatusProcessor {
+    private final GovernanceAggrProperties governanceAggrProperties;
     private final DRepStorage dRepStorage;
 
     private final List<DRepRegistrationEvent> dRepRegistrationsCache = Collections.synchronizedList(new ArrayList<>());
 
     @EventListener
     public void processDRepRegistrationEvent(DRepRegistrationEvent dRepRegistrationEvent) {
+        if (!governanceAggrProperties.isEnabled())
+            return;
+
         dRepRegistrationsCache.add(dRepRegistrationEvent);
     }
 
     @EventListener
     @Transactional
     public void handleCommitEventToProcessDRepStatus(PreCommitEvent preCommitEvent) {
+        if (!governanceAggrProperties.isEnabled())
+            return;
+
         try {
             dRepRegistrationsCache.sort(Comparator.comparingLong(e -> e.getMetadata().getSlot()));
             List<DRep> dReps = new ArrayList<>();
@@ -98,6 +106,9 @@ public class DRepStatusProcessor {
     @EventListener
     @Transactional
     public void handleRollbackEvent(RollbackEvent rollbackEvent) {
+        if (!governanceAggrProperties.isEnabled())
+            return;
+
         int count = dRepStorage.deleteBySlotGreaterThan(rollbackEvent.getRollbackTo().getSlot());
         log.info("Rollback -- {} drep records", count);
     }
